submissions-api/app/requirements.txt
# SPDX-FileCopyrightText: 2021 Genome Research Ltd.
#
# SPDX-License-Identifier: MIT

Cython == 0.29.21
connexion == 2.7.0
python_dateutil == 2.8.1
setuptools == 53.0.0
Werkzeug == 1.0.1
Flask-SQLAlchemy == 2.4.4
psycopg2-binary == 2.8.6
openpyxl == 3.0.6
jwt
et-xmlfile==1.0.1
minio==6.0.0
biopython==1.73

flask_testing==0.8.0
coverage>=4.0.3
pluggy>=0.3.1
py>=1.4.31
randomize>=0.13
responses
pytest==6.0.1